backend/requirements.txt
Flask
flask
psycopg2-binary
python-keycloak